# refactor-results/before-openapi.json
{
  "components": {
    "schemas": {
...
      "CaseDetail": {
        "additionalProperties": true,
        "properties": {
          "beneficiary_ids": {
            "items": {
              "type": "string"
            },
            "title": "Beneficiary Ids",
            "type": "array"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "notes": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Notes",
            "type": "array"
          },
          "risk_level": {
            "title": "Risk Level",
            "type": "string"
          },
          "risk_score": {
            "title": "Risk Score",
            "type": "integer"
          },
          "status": {
            "enum": [
              "Needs Review",
              "In Investigation",
              "Verification Requested",
              "Cleared"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "case_id",
          "beneficiary_ids",
          "risk_score",
          "risk_level",
          "status",
          "notes",
          "evidence"
        ],
        "title": "CaseDetail",
        "type": "object"
      },
      "CaseUpdate": {
        "properties": {
          "assigned_reviewer": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Reviewer"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 4000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "priority": {
            "anyOf": [
              {
                "enum": [
                  "Normal",
                  "High",
                  "Urgent"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "resolution": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "Needs Review",
                  "In Investigation",
                  "Verification Requested",
                  "Cleared"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "CaseUpdate",
        "type": "object"
      },
...
      "Evidence": {
        "properties": {
          "beneficiary_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beneficiary Id"
          },
          "contribution": {
            "title": "Contribution",
            "type": "integer"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "indicator": {
            "title": "Indicator",
            "type": "string"
          },
          "related_account_ids": {
            "items": {
              "type": "string"
            },
            "title": "Related Account Ids",
            "type": "array"
          },
          "related_beneficiary_ids": {
            "items": {
              "type": "string"
            },
            "title": "Related Beneficiary Ids",
            "type": "array"
          },
          "source_record_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Record Ids",
            "type": "array"
          }
        },
        "required": [
          "indicator",
          "contribution",
          "explanation",
          "related_beneficiary_ids",
          "related_account_ids",
          "source_record_ids"
        ],
        "title": "Evidence",
        "type": "object"
      },
      "FinancialFinding": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "assessment": {
            "enum": [
              "Review",
              "Context only"
            ],
            "title": "Assessment",
            "type": "string"
          },
          "beneficiary_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beneficiary Id"
          },
          "beneficiary_ids": {
            "items": {
              "type": "string"
            },
            "title": "Beneficiary Ids",
            "type": "array"
          },
          "case_ids": {
            "items": {
              "type": "string"
            },
            "title": "Case Ids",
            "type": "array"
          },
          "contribution": {
            "title": "Contribution",
            "type": "integer"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "finding_id": {
            "title": "Finding Id",
            "type": "string"
          },
          "indicator": {
            "title": "Indicator",
            "type": "string"
          },
          "related_account_ids": {
            "items": {
              "type": "string"
            },
            "title": "Related Account Ids",
            "type": "array"
          },
          "related_beneficiary_ids": {
            "items": {
              "type": "string"
            },
            "title": "Related Beneficiary Ids",
            "type": "array"
          },
          "source_record_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Record Ids",
            "type": "array"
          },
          "transactions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Transactions",
            "type": "array"
          }
        },
        "required": [
          "indicator",
          "contribution",
          "explanation",
          "related_beneficiary_ids",
          "related_account_ids",
          "source_record_ids",
          "finding_id",
          "beneficiary_ids",
          "case_ids",
          "transactions",
          "amount",
          "assessment"
        ],
        "title": "FinancialFinding",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "Page",
        "type": "object"
      },
      "ReportSnapshot": {
        "properties": {
          "anomalies": {
            "items": {
              "$ref": "#/components/schemas/FinancialFinding"
            },
            "title": "Anomalies",
            "type": "array"
          },
          "audit_id": {
            "title": "Audit Id",
            "type": "string"
          },
          "cases": {
            "items": {
              "$ref": "#/components/schemas/CaseDetail"
            },
            "title": "Cases",
            "type": "array"
          },
          "environment": {
            "title": "Environment",
            "type": "string"
          },
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "limitations": {
            "title": "Limitations",
            "type": "string"
          },
          "methodology": {
            "title": "Methodology",
            "type": "string"
          },
          "report_id": {
            "title": "Report Id",
            "type": "string"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          }
        },
        "required": [
          "report_id",
          "audit_id",
          "generated_at",
          "environment",
          "summary",
          "cases",
          "anomalies",
          "methodology",
          "limitations"
        ],
        "title": "ReportSnapshot",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local fictional scholarship audit prototype. Scores are review priorities, not fraud probabilities.",
    "title": "GrantLens",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cases": {
      "get": {
        "operationId": "cases_api_v1_cases_get",
        "parameters": [
          {
            "in": "query",
            "name": "audit_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Audit Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cases"
      }
    },
    "/api/v1/cases/{case_id}": {
      "get": {
        "operationId": "case_api_v1_cases__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "audit_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Audit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Case"
      },
      "patch": {
        "operationId": "update_case_api_v1_cases__case_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "audit_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Audit Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Case"
      }
    },
    "/api/v1/cases/{case_id}/report": {
      "get": {
        "operationId": "report_api_v1_cases__case_id__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "audit_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Audit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report"
      }
    },
...
    "/api/v1/dataset/sample/{filename}": {
      "get": {
        "operationId": "sample_file_api_v1_dataset_sample__filename__get",
        "parameters": [
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sample File"
      }
    },
...
    "/api/v1/reports": {
      "get": {
        "operationId": "reports_api_v1_reports_get",
        "parameters": [
          {
            "in": "query",
            "name": "audit_id",
            "required": true,
            "schema": {
              "title": "Audit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reports"
      },
      "post": {
        "operationId": "create_report_api_v1_reports_post",
        "parameters": [
          {
            "in": "query",
            "name": "audit_id",
            "required": true,
            "schema": {
              "title": "Audit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Report"
      }
    },
    "/api/v1/reports/{report_id}": {
      "get": {
        "operationId": "saved_report_api_v1_reports__report_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "audit_id",
            "required": true,
            "schema": {
              "title": "Audit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Saved Report"
      }
    },
    "/api/v1/reports/{report_id}/csv": {
      "get": {
        "operationId": "report_csv_api_v1_reports__report_id__csv_get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "audit_id",
            "required": true,
            "schema": {
              "title": "Audit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Archived case register CSV"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report Csv"
      }
    },
...
  }
}